refactor test mock tasks
